quantisation and buffering
added a fixed-point type to all the layers. This is coupled with an
"auto" quantisation method, which is 16-bit activation and weight
quantisation, where the binary point is infered for the weights from the
values. There's space for adding other quantisation methods as well.

added buffer depths to individual streams. This should be used with
buffer depth inference later on.